Cleaned up code
Mangled the code yesterday trying to connect to DB. Turns out the issue was my that my schema was capitalized. Mangled code was working, but I wanted the program to have uniformity so I went back and changed everything to it's original format. Seems to be working. Still need to change table names, but that is a quick fix.