Updated pom.xml and added Dockerfile
API Container created in Docker is giving error: Unable to locate main class java.ahmed.driver.Driver
However, this is the correct directory. I was thinking it may be possible that because the package is technically ahmed.Driver.Driver that this is the problem. If I change the directory for the main class in the jar-with-dependencies build Docker still give me the same error regarding java.ahmed.driver.Driver