Containers working
Containers working for now, however they will have to be deleted and recreated after code updates. Can login via HTML, however cannot access reimbursements via manager login, and changes are not entered into database when employee is logged in. This is because I do not have the correct address for the DB in fetch javascript. Need to figure out the correct format. Can access and make changes via Postman, but it will not work through HTML because I need the correct format in the logic folder for fetch line which is Javascript.